docs: reorganize Actions as separate *Action types
Adds the NavigateAction, EvaluateAction, KeyAction, MouseAction, and
QueryAction interfaces, all having the same signature as the original
Action interface.

Renames the MouseAction and KeyAction funcs to MouseEvent and KeyEvent, respectively.

Updates documentation on actions, and merges sel.go into query.go, as
well as sel_test.go into query_test.go. With the exception of some
outlier actions in nav.go, all the *Action are now nicely aligned in a
single <type>.go and <type>_test.go.

Additionally, eliminates unused TargetOption type.
